add trade mq consumer

+++ Pikachu-log-api/src/main/java/com/huiaong/pikachu/log/order/model/PikaPurchaseOrderOperationLog.java
@@ -15,5 +15,5 @@ public class PikaPurchaseOrderOperationLog extends PikaBaseBean {
     /**
      * @see PikaPurchaseOrderOperationLogType
      */
-    private Long operationType;
+    private Integer operationType;
 }
